tests/wpt/tests/css/filter-effects/parsing/flood-color-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: parsing flood-color with valid values</title>
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#FloodColorProperty">
<meta name="assert" content="flood-color supports the full grammar '<color>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("flood-color", "currentcolor");

test_valid_value("flood-color", "red");
test_valid_value("flood-color", "#00FF00", "rgb(0, 255, 0)");
test_valid_value("flood-color", "rgb(0, 0, 255)");
test_valid_value("flood-color", "rgb(100%, 100%, 0%)", "rgb(255, 255, 0)");
test_valid_value("flood-color", "hsl(120, 100%, 50%)", "rgb(0, 255, 0)");
test_valid_value("flood-color", "teal");

test_valid_value("flood-color", "transparent");
</script>
</body>
</html>
